third_party/WebKit/LayoutTests/media/track/track-webvtt-tc022-entities-expected.txt

Tests entities in the cue text.


*** Testing text track 0
EXPECTED (cues.length == '6') OK
EXPECTED (cues[0].getCueAsHTML().textContent == 'This cue has an amp & character.') OK
EXPECTED (cues[1].getCueAsHTML().textContent == 'This cue has a less than < character.') OK
EXPECTED (cues[2].getCueAsHTML().textContent == 'This cue has a greater than > character.') OK
EXPECTED (cues[3].getCueAsHTML().textContent == 'This cue has a Left-to-Right Mark ‎.') OK
EXPECTED (cues[4].getCueAsHTML().textContent == 'This cue has a Right-to-Left Mark ‏.') OK
EXPECTED (cues[5].getCueAsHTML().textContent == 'This cue has a non-breaking space  .') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].getCueAsHTML().textContent == 'This cue has an amp character. Ampersand is ignored.') OK
EXPECTED (cues[1].getCueAsHTML().textContent == 'This cue has a less than ') OK
EXPECTED (cues[2].getCueAsHTML().textContent == 'This cue has a greater than > character. Since it's not related to a < character, it's just interpreted as text.') OK
END OF TEST
